requirements.txt
matplotlib==3.7.2
numpy==1.25.0
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
plotly==5.15.0
plotly_express==0.4.1
Requests==2.31.0
SQLAlchemy==2.0.19
streamlit==1.24.0
streamlit_authenticator==0.2.2
streamlit_autorefresh==1.0.1
streamlit_option_menu==0.3.6
